🟢 filter test functions by endpoint
